Cleanup build process

+++ build-test.sh
@@ -2,14 +2,18 @@
 
 ACTION=${1:-}
 
-rm -r dist
+if [ -d dist ]; then
+    rm -r dist
+fi
 
 python3 -m build
 
 if [ "$ACTION" == "publish" ]; then
     python3 -m twine upload --repository testpypi dist/*
 
+    echo "Sleaping 30 secods to allow testpypi enough time to process upload of new version"
+    sleep 30s
+
     cd docker
-    docker build -f Dockerfile.test . --build-arg TEST_VERSION=1 --no-cache
-    cd ..
+    ./build.sh $ACTION
 fi
